+++ app/assets/stylesheets/modules/products/_products_new.scss
@@ -86,12 +86,12 @@
       }
     }
     .product__category__box{
-      @include product_box(100px);
+      padding:30px;
+      width:720px;
       margin-top: 130px;
       .product__details{
         @include product_name;
         color: rgb(204, 204, 204);
-        margin-bottom: 24px;
         .product__categoryField{
           @include product_fieldBox;
           &__name{
@@ -112,7 +112,6 @@
           &__selectarea__listen__added{
             select{
               margin-top: 16px;
-              margin-bottom: 14px;
               @include product_select;
             }
           }
@@ -121,7 +120,6 @@
     }
     .product__brand__box{
       @include product_box(100px);
-      margin-top: 50px;
       .product__brandField{
         @include product_fieldBox;
         &__name{
