+++ app/controllers/products_controller.rb
@@ -8,10 +8,9 @@ class ProductsController < ApplicationController
   def new
     @product = Product.new
     @product.images.new
+
     @category_parent_array = ["---"]
-    Category.where(ancestry: nil).each do |parent|
-      @category_parent_array << parent.name
-    end
+    @category_parent_array = Category.where(ancestry: nil)
     unless user_signed_in?
       redirect_to new_user_session_path
     end
@@ -27,6 +26,14 @@ class ProductsController < ApplicationController
     end
   end
 
+  def get_category_children
+    @category_children = Category.find("#{params[:parent_id]}").children
+  end
+
+  def get_category_grandchildren
+    @category_grandchildren = Category.find(params[:child_id]).children
+  end
+
   def show
   end
 
@@ -34,15 +41,6 @@ class ProductsController < ApplicationController
     flash.now[:alert] = '削除に失敗しました' unless @product.destroy  
   end
 
-
-  
-  def get_category_children
-    @category_children = Category.find_by(name: "#{params[:parent_name]}", ancestry: nil).children
-  end
-
-  def get_category_grandchildren
-    @category_grandchildren = Category.find_by(name: "#{params[:parent_name]}").children
-  end
 
   private
