+++ app/javascript/packs/application.js
@@ -10,6 +10,7 @@ require("channels")
 require('jquery')
 require('../products')
 require('../products_delete')
+require('../products_category')
 
 
 // Uncomment to copy all static images under ../images to the output folder and reference
