+++ app/javascript/products.js
@@ -82,94 +82,5 @@ $(window).on('load', function () {
   }
 });
 
-$(function(){
 
-  function appendOption(category){
-    var html = `<option value="${category.name}" datacategory="${category.id}">${category.name}</option>`;
-    return html;
-  }
 
-  function appendChildrenBox(insertHTML){
-    var childSelectHtml = '';
-    childSelectHtml = `<div class='category__styleSelect__selectarea__listen' id= 'children_wrapper'>
-                        <div class='category__styleSelect__selectarea__listen__added'>
-                          <select class="category__styleSelect__selectarea__listen__added__box" id="child_category" name="">
-                            <option value="---" data-category="---">---</option>
-                            ${insertHTML}
-                          </select>
-                        </div>
-                      </div>`;
-    $('.category__styleSelect__selectarea__listen').append(childSelectHtml);
-  }
-
-  function appendGrandchildrenBox(insertHTML){
-    var grandchildSelectHtml = '';
-    grandchildSelectHtml = `<div class='category__styleSelect__selectarea__listen' id= 'grandchildren_wrapper'>
-                              <div class='category__styleSelect__selectarea__listen__added'>
-                                <select class="lcategory__styleSelect__selectarea__listen__added__box" id="grandchild_category" name="category_id">
-                                  <option value="---" data-category="---">---</option>
-                                  ${insertHTML}
-                                </select>
-                              </div>
-                            </div>`;
-    $('.category__styleSelect__selectarea__two').append(grandchildSelectHtml);
-  }
-
-  $("#parent_category").on('change', function(){
-    var parentCategory = document.getElementById('parent_category').value;
-    if (parentCategory != "---"){
-      $.ajax({
-        url: 'get_category_children',
-        type: 'GET',
-        data: { parent_name: parentCategory },
-        dataType: 'json'
-      })
-      .done(function(children){
-        $('#children_wrapper').remove();
-        $('#grandchildren_wrapper').remove();
-        var insertHTML = '';
-        children.forEach(function(child){
-          insertHTML += appendOption(child);
-        });
-        appendChildrenBox(insertHTML);
-      })
-      .fail(function(){
-        alert('カテゴリー取得に失敗しました');
-      })
-    }else{
-      $('#children_wrapper').remove();
-      $('#grandchildren_wrapper').remove();
-    }
-  });
-
-  $(document).on('change', '#child_category', function(){
-    var parentCategory = document.getElementById('child_category').value;
-    if (parentCategory != "---"){
-      $.ajax({
-        url: 'get_category_grandchildren',
-        type: 'GET',
-        data: { parent_name: parentCategory },
-        dataType: 'json'
-      })
-      .done(function(grandchildren){
-        if(grandchildren.length != 0) {
-          $('#grandchildren_wrapper').remove();
-          $('#size_wrapper').remove();
-          $('#brand_wrapper').remove();
-          var insertHTML = '';
-          grandchildren.forEach(function(grandchild){
-            insertHTML += appendOption(grandchild);
-          });
-          appendGrandchildrenBox(insertHTML);
-        }
-      })
-      .fail(function(){
-        alert('カテゴリー取得に失敗しました');
-      })
-    }else{
-      $('#grandchildren_wrapper').remove();
-      $('#size_wrapper').remove();
-      $('#brand_wrapper').remove();
-    }
-  });
-});
